fix(calibrate): fix incorrect cropping on iOS
On iOS innerWidth and clientWidth give wrong results when called
at that point of time. Mysteriously it's always '980' no matter
which device.

window.screen.width gives the correct value, so use that on iOS.

Having the correct width here will also result in using the
correct pixel ratio later on. And with the correct pixel ratio
screenshots on iOS are not cropped incorrectly, which was the
original motivation for this fix.